requirements.txt
discord.py
asyncio
requests
DiscordUtils
motor
pathlib
buttons
